Don't require numfig to use :numref: on sections
Tweak logic that rejects a :numref: if numfig is not on to bypass this
check if the reference is a section. Section numbers are applied
independent of numfig, so the check is not needed, and makes it
needlessly difficult to use :numref: if the user only cares about using
it on sections.