Update the screenshots and docs to match the Python website and installer
The instructions for installing pip are no longer relevant and have
been removed. The instructions for modifying the PATH are no longer
relevant and have been removed. The final section, "Installing Sphinx
with pip", is now a subsection of the Windows install instructions.
The verb tense has been modified to match the rest of the instructions.
The command line examples have been converted to use the Pygments text
lexer instead of the batch lexer, which highlights everything after `>`
incorrectly.