Implement delayed resolution in TOC
Implement a new node that converts its contents to text (via astext()),
but delays doing so until the post-transform stage. Modify TOC
collection to use this. Modify TOC extraction to perform post-transforms
(specifically, reference resolution and resolution of the aforementioned
new nodes).

This fixes an annoying behavior where the use of an external reference
in a section name would use the raw reference text rather than the
resolved reference text in the sidebar table of contents, resulting in a
mismatch between the section title as it appears in said table of
contents and how it appears in the document body.

Fixes #4448.