Allow imgcls as class for SVG Graphviz <object>
For consistency, the `<object>` tag for the SVG output of a
inheritance-diagram should have the same `class="inheritance"` attribute
that the corresponding `<img>` tag for the PNG format has. This allows
to define the CSS attribute

    object.inheritance {
        max-width: 100%;
    }

which for SVG (unlike PNG) actually works and makes sense.

Added a test for inheritance-diagrams in SVG.