LaTeX: add xindy merge rules for TeX-escaped Greek letters
The effect is that indexed Greek letters are sorted with uppercase Greek
letters before lowercase Greek: ΓΔΘ...αβγ...; without this commit the
Greek letters are ordered according to the ascii characters used in the
TeX control sequences as set-up in sphinx.util.texescape, i.e. the order
is αβχΔδεηΓγ... In both cases indexed terms starting with such a Greek
letter end up in the same group as terms starting with a non-letter.

This commit does not create letter-groups for Greek letters, as such
letter-groups would require to make sure non-escaped letters such as α
are accepted by TeX engine, and this requires extra-steps for pdflatex (or
OpenType font with Greek script for xelatex/lualatex), and if those
steps were made default by Sphinx, there would not be a reason anymore
to escape via sphinx.util.texescape the Greek letters from rest source.