Revert "Builder.build: save env also with only new documents"
This reverts commit 20f625b4d39d0e981b4922f629212acc74271665, which
introduced a regression in dependency tracking for build systems
invoking sphinx-build (described in issue #5501.)

The correct solution is to properly track when the pickle file needs to
be written;  unfortunately I don't know enough about sphinx to implement
this.  Also, I'm not sure the intersphinx cache belongs into this at
all.