Fix `options` decoration and implement `objtype` aliases
To check whether something is a method, it has not been convenient to
check the objtype, because it could be `method`, `staticmethod`, or
`classmethod`.

Since the behavioural change associated to the last two object types
is now provided by setting the `static` or `classmethod` flags,
respectively, we can take the aliasing to the next level and instruct
`objtype` to reveal the “true” objtype: `staticmethod` is now aliased
to `method`.

The implementation of the `options` decoration had the serious flaw
that the dict was essentially immutable because the getter always
returned a copy of the stored version.

This is fixed by initializing it to an `AliasedDict` which overrides
`__init__` and `__setitem__` of `dict` to target the aliased key
instead (e.g. `abstract` instead of `abstractmethod`).