Add support for booktabs-style tables to LaTeX builder
Render tables without vertical rules and horizontal rules of varying
thickness (with additional space above and below) using the booktabs
package.

This is a rebase of original commit which was on master branch prior to
2.1 release.  In this rebase the \RequirePackage{booktabs} has been
removed from sphinx.sty, as it has to be conditional on latex_booktabs
(or whatever its name will be) configuration option, and will be
repositioned in a subsequent commit.