Add get_members function to apidoc templates
The `module.rst_t` and `package.rst_t` templates used by apidoc now have
a `get_members` function that allows to obtain the member of the current
module or package in various forms and filtered by "type" (exception,
class, function, data).

This provides a powerful mechanism for writing more advanced templates.
Specifically, it allows to generate structured summaries of the contents
of a module, preceding the detailed `automodule` references.
See https://krotov.readthedocs.io/en/stable/API/krotov.html for an
example.

Extensive documentation of all the templating capabilities is included
in the apidoc man page in the documentation.

This functionality is a backport of the "better-apidoc" package
(https://github.com/goerz/better-apidoc).