LaTeX: fix #6804
In case a code-block is in an environment already using the "framed"
package, measure the vertical size needed for the code lines and use a
nested "framed" only if short enough (i.e. at most 90% of text area on a
page).

Long nested contents will drop the inner frame. They lose the background
color as well as the continuation hints across pagebreaks. But the
caption if present is typeset normally. And highlighted lines and line
numbers are rendered with no alteration.

Short nested contents (should) render exactly as before this commit:
they use "framed" but this must be encapsulated in an unbreakable
"minipage". All decoration including caption are obeyed but of course
not the continuation hints are these never apply.