LaTeX: revert in index entries to \textbar{} not \sphinxtextbar{}
This is because of case of an index directive or role in a table cell.
For the same reason tabulary is not compatible with code-cells,
it affects index entries. The escaping of | char by a LaTeX macro
needed to be by a "robus" or "protected" one, or by string like {"|}.
This was not the choice of previous commit, by oversight.

By reverting to \textbar{} (which is a "robust" command) we also
maintain maximal backwards compatibility for the sorting in the
index done by the makeindex program.