sphinx-build: Guess sourcedir if none provided
This is part two of two of the "make 'sphinx-build' easier to call"
effort. In this part, we attempt to guess where the source directory is
if the user does not provide one. This is possible because many users
use one of three directory names - doc, docs, or Documentation - for
their documentation. We check these, along with all immediate
subdirectories for these directories. If a 'conf.py' file is found in
any of these directories, we assume this is the documentation source.
This means a user no longer needs to pass any argument to 'sphinx-build'
is they do not wish to.

This is, of course, best-effort, and users can continue to provide these
options manually if they so wish.

Signed-off-by: Stephen Finucane <stephen@that.guru>