Use path2doc rather than reinventing the wheel
We already have a helper function that allows us to convert a filename
to a "docname". Don't reinvent the wheel and simply use this when
building specific files with Sphinx. Note that this does change behavior
slightly, insofar as filenames that don't resolve to valid docnames will
now be ignored rather than being passed with their suffix and silently
ignored later, but that seems sane.

Signed-off-by: Stephen Finucane <stephen@that.guru>