docs: Split builders page into multiple pages
We want to move more documents from the root document tree into the user
guide. This requires breaking our builder document into multiple
documents, each focused on a different output format.

Note that, for the most part, the content is unchanged and is merely
shuffled around. This is intentional since we can address additional
things easily in later changes once the basic structure is in place.

Signed-off-by: Stephen Finucane <stephen@that.guru>