Ignore bdb.BdbQuit when handling exceptions
`bdb.BdbQuit` is used when quitting the debugger.
It should not a) cause the debugger to be started (with `-P` / `--pdb`),
and b) not a "crash" to be logged.

This helps when using `pdb.set_trace()` manually, and quitting it with
`q`.

It gets used in `build_main`, and `BuildDoc.run` (distutils command).