Configuration of pointer/ref placement in C++ decls
This change adds options 'cpp_ptr_placement' and 'cpp_ref_placement',
which determine where the rendered output will attach the reference
`&` or pointer `*` declaration token relative to the main type and
the declared name. "right" (the current behavior and default) will
attach towards the right on the declared name, whereas "left" will
attach towrads the left, closer to the pointed-to/referred type.