refactor: autodoc: if-block for callable object is no longer needed
The if-block for callable object was added to FunctionDocumenter at
5c3a0e4e40 (see #5447).  At that time, it helps to introspect
the function signature for a callable object because
``sphinx.util.inspect:Signature`` emits a warning when a callable
object is given.

For now, ``sphinx.util.inspect:signature()` does not emit a warning
even if a callable object is given.  So this if-block is no longer
needed.