docs: Remove the 'develop' guide
This contained a list of extensions, along with general contributing
information and pointers on how to develop your own extensions. The list
of extensions was very old, referenced the dead BitBucket repo and was
really not very helpful, while the contributing and development
information was just pointers to other, better documentation. Remove the
document wholesale and update any pointers we have to it to other,
better docs.

Signed-off-by: Stephen Finucane <stephen@that.guru>