Fix #7802: autodoc: EOFError is raised on parallel build
sphinx-build has crashed with EOFError when autodoc raises a warning
having exc_info under parallel mode.  In parallel mode, all messages
are pickled to transfer logs to parent process.  But the warning is
not picklable because it contains a traceback object.

This removes exc_info from warning messages to prevent crashes.