Fixes #8146: When identifying bases, only use classes from builtins
In inheritance_diagram extension, while iterating over bases,
we verify if the base class is one of the Python built-in
class or not.
As of now, we simply check for its presence in all `builtins`
objects. Please note, `builtins` not only has built-in classes,
but also functions (like `open`) and other built-in objects.

To avoid any sort of future problem, it seems better to only
use classes (and of course exception classes).