Fix 'oneside' latex output
Sphinx.sty had an FIXME bug where documents with the 'oneside' option
would output lots of the following warnings when building the document:

   Package Fancyhdr Warning: \fancyfoot's `E' option without twoside
   option is useless on input line 309.

Fix the problem by using \if@twoside around the offending tex code that
tries to set format for the even pages. 'Even' pages don't exist on
oneside builds, which causes the warning to be emitted.

Fixes: #8215