Strip the token name from productionlist.
When the user tries to align production rules, Sphinx gets confused by
spaces around leading tokens. In the following, references to mygroup:bad1
and mygroup:bad2 are missing, while the one to mygroup:correct is present.

    .. productionlist:: mygroup
       correct: `correct` | `bad1` | `bad2`
       bad1   : `correct` | `bad1` | `bad2`
          bad2: `correct` | `bad1` | `bad2`

Since it is useful to be able to align colons in a production list, this
commit systematically strips leading and trailing whitespaces from the
production name.