domain: c: fix macro usage inside struct
This fixes issues seen when a macro is used inside a struct. A simple
example:

```
/** @brief some struct */
typedef struct {
    /** Some macro */
    SOME_MACRO(a, b);
} t;
```

If `SOME_MACRO` was declared outside of the file, Doxygen generates a
function element. There current c-domain routines do not handle untyped
functions, so this commit adds an extra check when a typed function
parsing generates a DefinitionError to also try to parse it as an
untyped function.

Signed-off-by: Fabio Utzig <fabio.utzig@nordicsemi.no>